refractor: dates functions -> datesUtil
