+++ frontend/src/components/Feature/History/HistoryManageTable.tsx
@@ -6,7 +6,7 @@ import React from 'react';
 import { HistorySnapshotBundledData, HistorySnapshotResponse, MakeActiveSnapshotRequestDto, TrackResponse } from '../../../util/types';
 import { useCustomAlert } from '../../Hooks/useCustomAlert';
 import { mapTemplateInputsToTemplateViewerFromSnapshot, templateInputsFromHistorySnapshot } from '../../../util/dataHelpers';
-import { delay, getPlugins, utcFormatter } from '../../../util/helper';
+import { delay, getPlugins } from '../../../util/helper';
 import { SpinnerDefault } from '../../MenuItems/Loaders/SpinnerDefault';
 import { ConfirmationPopup } from '../../MenuItems/Popups/ConfirmationPopup';
 import { AlertPopup, CustomAlertPopupProps, PopupType } from '../../MenuItems/Popups/AlertPopup';
@@ -19,7 +19,8 @@ import { SaveButton } from '../../MenuItems/Buttons/SaveButton';
 import { VerifyIcon } from '../../MenuItems/Icons/VerifyIcon';
 import { useCustomConfirmationPopup } from '../../Hooks/useCustomConfirmationPopup';
 import { CloseWindowIcon } from '../../MenuItems/Icons/CloseWindowIcon';
-import { getFontsData } from '../../../util/fontsUtility';
+import { getFontsData } from '../../../util/fontsUtil';
+import { utcFormatter } from '../../../util/datesUtil';
 
 type Props = {
     getHistory: () => void;
